QMAPS-1883 move return button logic in panelManager
