+++ views/index.ejs
@@ -49,7 +49,7 @@
         ></button>
         <div class="search_form__wrapper empty">
           <div class="search_form__return icon-arrow-left"></div>
-          <div class="search_form__return-to-list icon-arrow-left" onclick="window.history.go(-1); return false;"></div>
+          <div class="search_form__return-to-list icon-arrow-left"></div>
           <input id="search" type="search" class="search_form__input" spellcheck="false" required placeholder="<%= _('Search on Qwant Maps') %>" autocomplete="off">
           <button id="clear_button_mobile" class="search_form__clear icon-x" type="button" onmousedown="clearSearch(event)" title="<%= _('Clear', 'search bar') %>"></button>
           <input type="submit" value="" class="search_form__action" onclick="submitSearch()" title="<%= _('Search') %>">
